Move some data-related code to new service
