+++ apps/maptio/src/app/workspace/workspace.component.ts
@@ -6,22 +6,16 @@ import {
   ChangeDetectionStrategy,
   inject,
 } from '@angular/core';
-import { NgClass } from '@angular/common';
 import { ActivatedRoute } from '@angular/router';
 
 import { Subscription } from 'rxjs';
 import { tap } from 'rxjs/operators';
 
-import { Store } from '@ngrx/store';
-
 import { SidePanelLayoutComponent } from '@notebits/toolkit';
 
-import { EmitterService } from '@maptio-core/services/emitter.service';
 import { DataSet } from '@maptio-shared/model/dataset.data';
 import { Team } from '@maptio-shared/model/team.data';
 import { User } from '@maptio-shared/model/user.data';
-import { AppState } from '@maptio-state/app.state';
-import { setCurrentOrganisationId } from '@maptio-state/global.actions';
 
 import { WorkspaceService } from '@maptio-workspace/workspace.service';
 import { MapContainerComponent } from '@maptio-workspace/map-container/map-container.component';
@@ -31,25 +25,22 @@ import { OnboardingVideoComponent } from './onboarding-video/onboarding-video.co
 
 // TODO: Move these components slowly into the new standalone component workspace
 import { WorkspaceFacade } from '@maptio-old-workspace/+state/workspace.facade';
-import { BuildingComponent } from '@maptio-old-workspace/components/data-entry/hierarchy/building.component';
 import { UIService } from '@maptio-old-workspace/services/ui.service';
 import { MapSettingsService } from '@maptio-old-workspace/services/map-settings.service';
 
 @Component({
-    selector: 'maptio-workspace',
-    templateUrl: 'workspace.component.html',
-    styleUrls: ['./workspace.component.scss'],
-    changeDetection: ChangeDetectionStrategy.OnPush,
-    imports: [
-    BuildingComponent,
-    NgClass,
+  selector: 'maptio-workspace',
+  templateUrl: 'workspace.component.html',
+  styleUrls: ['./workspace.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  imports: [
     SidePanelLayoutComponent,
     MapContainerComponent,
     StructureEditorContainerComponent,
     InitiativeDetailsContainerComponent,
-    OnboardingVideoComponent
-],
-    providers: [UIService, MapSettingsService, WorkspaceFacade]
+    OnboardingVideoComponent,
+  ],
+  providers: [UIService, MapSettingsService, WorkspaceFacade],
 })
 export class WorkspaceComponent implements OnInit, OnDestroy {
   workspaceService = inject(WorkspaceService);
@@ -59,7 +50,6 @@ export class WorkspaceComponent implements OnInit, OnDestroy {
   constructor(
     private route: ActivatedRoute,
     private cd: ChangeDetectorRef,
-    private store: Store<AppState>
   ) {}
 
   ngOnInit() {
@@ -107,51 +97,22 @@ export class WorkspaceComponent implements OnInit, OnDestroy {
                   this.workspaceService.isLoading.set(false);
                   this.cd.markForCheck();
                 });
-            }
-          )
+            },
+          ),
         )
         .subscribe(
           (data: {
             data: { dataset: DataSet; team: Team; members: User[]; user: User };
           }) => {
-            this.workspaceService.dataset.set(data.data.dataset);
-            this.workspaceService.tags.set(data.data.dataset.tags);
-            this.workspaceService.team.set(data.data.team);
-            this.workspaceService.members.set(data.data.members);
-            this.workspaceService.user.set(data.data.user);
-            this.workspaceService.datasetId.set(
-              this.workspaceService.dataset().datasetId
-            );
-            this.workspaceService.teamName.set(
-              this.workspaceService.team().name
-            );
-            this.workspaceService.teamId.set(
-              this.workspaceService.team().team_id
-            );
-            EmitterService.get('currentTeam').emit(
-              this.workspaceService.team()
-            );
-
-            const currentOrganisationId = this.workspaceService.team()?.team_id;
-            this.store.dispatch(
-              setCurrentOrganisationId({ currentOrganisationId })
-            );
-
-            this.workspaceService.isEmptyMap.set(
-              !this.workspaceService.dataset().initiative.children ||
-                this.workspaceService.dataset().initiative.children.length === 0
-            );
+            this.workspaceService.loadData(data.data);
             this.cd.markForCheck();
-          }
+          },
         );
     });
   }
 
   ngOnDestroy(): void {
-    EmitterService.get('currentTeam').emit(undefined);
-    this.store.dispatch(
-      setCurrentOrganisationId({ currentOrganisationId: undefined })
-    );
+    this.workspaceService.clearGlobalStateOnDestroy();
 
     if (this.routeSubscription) this.routeSubscription.unsubscribe();
   }
